Add autoscaling for addon (#45)
Picking the avg_task_duration is a bit tricky. The median job time varies between devedition and firefox for 70.0b12 - 1m 0s vs 4m40s - as if they are doing different amount of work but the same # of xpi are being submitted. I've tended towards the Firefox value by taking 4m. A failing job runs out of retries after ~13 minutes.

We have 4 instances at the moment, and the l10n chunks tend to finish in a fairly small time window, so a max_replicas of 10 should mean higher throughput. We shall to see if amo responds ok to that.